requirements.txt
aiomysql==0.0.22
aioredis==1.3.1
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
atlastk==0.12.11
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
fastapi==0.70.0
h11==0.12.0
hiredis==2.0.0
idna==3.3
iso8601==0.1.16
jwt==1.3.1
pycparser==2.21
pydantic==1.8.2
PyMySQL==0.9.3
pypika-tortoise==0.1.1
pytz==2021.3
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing_extensions==4.0.1
uvicorn==0.15.0
